Add some information to begin with
Added the scope of the talk, which is by now just
temporary. I sketched some ideas of what I picture
the talk to be, in a <Pre-research> section.

I included the definition of generative art, given
by Phillip Galanter 2003. Be sure to add cites to this.

Added ideas to be coding for the moment of the talk:
  - Pixel sort